\name{getTimestamp}
\alias{getTimestamp}
\title{
getTimestamp
}
\description{
Returns the current timestamp.
}
\usage{
getTimestamp()
}
\value{
numeric vector with 2 values (number of seconds since epoch, and number of microseconds in current second).
}
\author{
Pierrick Bruneau
}
\seealso{
getElapsed
}
\examples{
stamp <- getTimestamp()
}

