% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablefile.R
\name{isTsv}
\alias{isTsv}
\title{determine file format from file name}
\usage{
isTsv(path)
}
\arguments{
\item{path}{tsv: \code{*.txt}, \code{*.xls}; csv: \code{*.csv}}
}
\description{
determine file format from file name
}
