% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{parseSize}
\alias{parseSize}
\title{parse size value}
\usage{
parseSize(value, default = list(w = 1400, h = 1200))
}
\arguments{
\item{value}{value can be character, numeric and list}

\item{default}{the default size value is the parser test failure on the
value parameter.}
}
\value{
a list object that contains fields \code{w} and \code{h}.
}
\description{
parse size value
}
