% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablefile.R
\name{write_table.auto}
\alias{write_table.auto}
\title{Write table safely}
\usage{
write_table.auto(data, path)
}
\arguments{
\item{data}{a \code{data.frame} object or matrix data object.}

\item{path}{create a file safely, create directory if the parent
directory is not exists, and determine file format by the file
name its extension name.}
}
\description{
Write table safely
}
